// src/services/api/guests.types.ts
export type RawDataGuest = {
  id: number;
  country_flag: string | null;
  email: string | null;
  full_name: string | null;
  national_id: string | null;
  nationality: string | null;
};

export type DataGuest = {
  id: number;
  countryFlag: string | null;
  email: string | null;
  fullName: string | null;
  nationalId: string | null;
  nationality: string | null;
};